requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
bcrypt
aioredis
asyncio
aiohttp
openai
googlemaps
stripe
redis
xgboost
joblib
numpy
scikit-learn
psutil
python-dotenv
pytest
fastapi-limiter
google-cloud-dialogflow
googletrans